requirements.txt
python-telegram-bot == 12.7
babel == 2.7.0
sqlalchemy == 1.3.16
alembic == 1.4.2
flexiconf == 1.0.4
